+++ Models/manyToMany/ContentSession.cs
@@ -9,11 +9,11 @@ public partial class ContentSession
     {
     }
 
-    public long ProfileId {get; set;}
-    public long EpisodeId {get; set;}
+    public int ProfileId {get; set;}
+    public int EpisodeId {get; set;}
 
     public bool Active {get; set;}
-    public long LeftOffAt {get; set;}
-    public long TimeWatched {get; set;}
+    public int LeftOffAt {get; set;}
+    public int TimeWatched {get; set;}
 
 }
